backend_fastapi/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic>=2.7.4
pydantic-settings>=2.4.0

# Database
mysql-connector-python==8.3.0
pymongo==4.6.1
motor==3.3.2

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Email
# aiosmtp==0.1.2  # Package not available, using aiosmtplib instead
aiosmtplib==3.0.0
email-validator==2.1.0.post1

# HTTP Requests
httpx==0.26.0
aiofiles==23.2.1

# AI/ML
openai==1.10.0
groq>=0.29.0

# Web Scraping (for PMDC verification)
selenium==4.17.2
beautifulsoup4==4.12.3

# Image Processing (for future CNN)
Pillow==10.2.0
numpy==1.26.3

# Validation
email-validator==2.1.0.post1
